datasets files

+++ Entity Framework Core/JavaScript Object Notation - JSON/CarDealer/CarDealer.Models/Car.cs
@@ -12,7 +12,7 @@ namespace CarDealer.Models
         public int Id { get; set; }
         public string Make { get; set; }
         public string Model { get; set; }
-        public int TravelledDistance { get; set; }
+        public long TravelledDistance { get; set; }
 
         public virtual ICollection<PartCar> CarParts { get; set; }
     }
